+++ ui/dev/src/pages/TestSignaturePad.vue
@@ -12,7 +12,7 @@
       <qas-signature-uploader v-model="signature" class="col-12" entity="portal-app/problems/photo" signature-label="Assinatura do Cleyton" upload-label="Assinatura digital" />
       <pre class="q-mb-xl">v-model {{ signature }}</pre>
 
-      <qas-form-generator v-model="formGenerator.values" columns="fit" :errors="formGenerator.errors" :fields="formGenerator.fields" />
+      <qas-form-generator v-model="formGenerator.values" columns="fit" :errors="formGenerator.errors" :fields="formGenerator.fields" :fields-props="filesFieldsProps" />
       <pre>{{ formGenerator }}</pre>
     </div>
   </q-page>
@@ -47,6 +47,14 @@ export default {
             'Campo obrigatório'
           ]
         }
+      }
+    }
+  },
+
+  computed: {
+    filesFieldsProps () {
+      return {
+        uploader: { readonly: true }
       }
     }
   },